+++ __tests__/server/mixins.js
@@ -6,31 +6,26 @@ const mixins = require('../../src/server/mixins')
 describe('mixins', () => {
   let db
 
-  before(() => {
+  beforeAll(() => {
     _.mixin(lodashId)
     _.mixin(mixins)
   })
 
   beforeEach(() => {
     db = {
-      posts: [
-        { id: 1, comment: 1 }
-      ],
+      posts: [{ id: 1, comment: 1 }],
       comments: [
         { id: 1, postId: 1 },
         // Comments below references a post that doesn't exist
         { id: 2, postId: 2 },
         { id: 3, postId: 2 }
       ],
-      photos: [
-        { id: '1' },
-        { id: '2' }
-      ]
+      photos: [{ id: '1' }, { id: '2' }]
     }
   })
 
   describe('getRemovable', () => {
-    it('should return removable documents', () => {
+    test('should return removable documents', () => {
       const expected = [
         { name: 'comments', id: 2 },
         { name: 'comments', id: 3 }
@@ -39,7 +34,7 @@ describe('mixins', () => {
       assert.deepEqual(_.getRemovable(db, { foreignKeySuffix: 'Id' }), expected)
     })
 
-    it('should support custom foreignKeySuffix', () => {
+    test('should support custom foreignKeySuffix', () => {
       const expected = [
         { name: 'comments', id: 2 },
         { name: 'comments', id: 3 }
@@ -50,11 +45,11 @@ describe('mixins', () => {
   })
 
   describe('createId', () => {
-    it('should return a new id', () => {
+    test('should return a new id', () => {
       assert.equal(_.createId(db.comments), 4)
     })
 
-    it('should return a new uuid', () => {
+    test('should return a new uuid', () => {
       assert.notEqual(_.createId(db.photos), 3)
     })
   })