+++ __tests__/server/singular.js
@@ -21,16 +21,15 @@ describe('Server', () => {
   })
 
   describe('GET /:resource', () => {
-    it('should respond with corresponding resource', () => (
+    test('should respond with corresponding resource', () =>
       request(server)
         .get('/user')
         .expect(db.user)
-        .expect(200)
-    ))
+        .expect(200))
   })
 
   describe('POST /:resource', () => {
-    it('should create resource', () => {
+    test('should create resource', () => {
       const user = { name: 'bar' }
       return request(server)
         .post('/user')
@@ -41,7 +40,7 @@ describe('Server', () => {
   })
 
   describe('PUT /:resource', () => {
-    it('should update resource', () => {
+    test('should update resource', () => {
       const user = { name: 'bar' }
       return request(server)
         .put('/user')
@@ -52,12 +51,11 @@ describe('Server', () => {
   })
 
   describe('PATCH /:resource', () => {
-    it('should update resource', () => (
+    test('should update resource', () =>
       request(server)
         .patch('/user')
         .send({ name: 'bar' })
         .expect({ name: 'bar', email: 'foo@example.com' })
-        .expect(200)
-    ))
+        .expect(200))
   })
 })