+++ src/cli/utils/is.js
@@ -1,17 +1,17 @@
 module.exports = {
-  JSON,
+  FILE,
   JS,
   URL
 }
 
-function JSON (s) {
+function FILE(s) {
   return !URL(s) && /\.json$/.test(s)
 }
 
-function JS (s) {
+function JS(s) {
   return !URL(s) && /\.js$/.test(s)
 }
 
-function URL (s) {
+function URL(s) {
   return /^(http|https):/.test(s)
 }