+++ src/server/defaults.js
@@ -6,13 +6,12 @@ const cors = require('cors')
 const compression = require('compression')
 const errorhandler = require('errorhandler')
 const objectAssign = require('object-assign')
+const bodyParser = require('./body-parser')
 
-module.exports = function (opts) {
+module.exports = function(opts) {
   const userDir = path.join(process.cwd(), 'public')
-  const defaultDir = path.join(__dirname, 'public')
-  const staticDir = fs.existsSync(userDir)
-    ? userDir
-    : defaultDir
+  const defaultDir = path.join(__dirname, '../../dist')
+  const staticDir = fs.existsSync(userDir) ? userDir : defaultDir
 
   opts = objectAssign({ logger: true, static: staticDir }, opts)
 
@@ -21,18 +20,6 @@ module.exports = function (opts) {
   // Compress all requests
   if (!opts.noGzip) {
     arr.push(compression())
-  }
-
-  // Logger
-  if (opts.logger) {
-    arr.push(
-      logger('dev', {
-        skip: (req) => (
-          process.env.NODE_ENV === 'test' ||
-          req.path === '/favicon.ico'
-        )
-      })
-    )
   }
 
   // Enable CORS for all the requests, including static files
@@ -47,6 +34,16 @@ module.exports = function (opts) {
 
   // Serve static files
   arr.push(express.static(opts.static))
+
+  // Logger
+  if (opts.logger) {
+    arr.push(
+      logger('dev', {
+        skip: req =>
+          process.env.NODE_ENV === 'test' || req.path === '/favicon.ico'
+      })
+    )
+  }
 
   // No cache for IE
   // https://support.microsoft.com/en-us/kb/234067
@@ -66,6 +63,11 @@ module.exports = function (opts) {
         res.sendStatus(403) // Forbidden
       }
     })
+  }
+
+  // Add middlewares
+  if (opts.bodyParser) {
+    arr.push(bodyParser)
   }
 
   return arr