+++ src/server/router/nested.js
@@ -1,11 +1,13 @@
 const express = require('express')
 const pluralize = require('pluralize')
+const delay = require('./delay')
 
-module.exports = (opts) => {
+module.exports = opts => {
   const router = express.Router()
+  router.use(delay)
 
   // Rewrite URL (/:resource/:id/:nested -> /:nested) and request query
-  function get (req, res, next) {
+  function get(req, res, next) {
     const prop = pluralize.singular(req.params.resource)
     req.query[`${prop}${opts.foreignKeySuffix}`] = req.params.id
     req.url = `/${req.params.nested}`
@@ -13,7 +15,7 @@ module.exports = (opts) => {
   }
 
   // Rewrite URL (/:resource/:id/:nested -> /:nested) and request body
-  function post (req, res, next) {
+  function post(req, res, next) {
     const prop = pluralize.singular(req.params.resource)
     req.body[`${prop}${opts.foreignKeySuffix}`] = req.params.id
     req.url = `/${req.params.nested}`